+++ Elements.MEP/src/Fittings/FlowDirection.cs
@@ -1,0 +1,8 @@
+namespace Elements.Fittings
+{
+    public enum FlowDirection : byte
+    {
+        TowardLeafs,
+        TowardTrunk
+    }
+}