+++ Elements/src/Geometry/ThickenedPolyline.cs
@@ -299,14 +299,18 @@ namespace Elements.Geometry
             // Populate the offset vertex map for each node in the graph. These points can be used to construct the thickened geometry.
             PopulateOffsetVertexMap(graph, normalDir);
 
-            var polygons = new List<Polygon>();
-            var lines = new List<Line>();
             // For each edge in the graph, construct a new polygon from the points in the offset vertex map.
             foreach (var (a, b, origPlIndex) in graph.Edges)
             {
 
                 var abc = graph.Nodes[a].offsetVertexMap[b];
                 var def = graph.Nodes[b].offsetVertexMap[a];
+                // if the edge has a thickness of 0, we may get either collinear vertices or fully collapsed vertices. In either case, we can just return a line segment.
+                if (abc.Union(def).ToList().AreCollinearByDistance() || Vector3.AreApproximatelyEqual(abc) || Vector3.AreApproximatelyEqual(def))
+                {
+                    resultList[origPlIndex] = (offsetPolygon: null, offsetLine: new Line(abc[1], def[1]));
+                    continue;
+                }
                 try
                 {
                     var pgonOutput = new Polygon(abc.Concat(def).ToArray());
@@ -323,11 +327,14 @@ namespace Elements.Geometry
                         // get largest offset
                         var largestOffset = offsets.OrderByDescending((pgon) => Math.Abs(pgon.Area())).First();
                         resultList[origPlIndex] = (offsetPolygon: largestOffset.Perimeter, offsetLine: null);
-                        Elements.Validators.Validator.DisableValidationOnConstruction = false;
                     }
                     catch
                     {
                         resultList[origPlIndex] = (offsetPolygon: null, offsetLine: new Line(abc[1], def[1]));
+                    }
+                    finally
+                    {
+                        Elements.Validators.Validator.DisableValidationOnConstruction = false;
                     }
                 }
 
@@ -345,6 +352,52 @@ namespace Elements.Geometry
         public List<(Polygon offsetPolygon, Line offsetLine)> GetPolygons(Vector3? normal = null)
         {
             return GetPolygons(new[] { this }, normal);
+        }
+
+        /// <summary>
+        /// Given an ordered list of vertices representing a loop, and a
+        /// corresponding list of thicknesses for each edge, construct the
+        /// polygonal boundary taking into account those thickness offsets.
+        /// </summary>
+        /// <param name="vertices">An ordered list of vertices representing a closed polygon.</param>
+        /// <param name="edgeThickness">An ordered list of leftWidth/rightWidth pairs, representing the thickness of the edge beginning at each vertex.</param>
+        /// <param name="left">Which side of the loop boundary to compute. Left by default.</param>
+        /// <param name="normal">The normal defining the plane in which to calculate the boundary.</param>
+        public static Polygon GetLoopBoundary(IList<Vector3> vertices, IList<double[]> edgeThickness, bool left = true, Vector3? normal = null)
+        {
+            var normalDir = normal ?? Vector3.ZAxis;
+            if (vertices.Count != edgeThickness.Count)
+            {
+                throw new ArgumentException("The number of vertices must match the number of edge thicknesses.");
+            }
+            var polylines = new List<ThickenedPolyline>();
+            for (int i = 0; i < vertices.Count; i++)
+            {
+                var vertex = vertices[i];
+                var nextVertex = vertices[(i + 1) % vertices.Count];
+                var thickness = edgeThickness[i];
+                polylines.Add(new ThickenedPolyline(new[] { vertex, nextVertex }, thickness[0], thickness[1]));
+            }
+            // Initialize a graph to manage nodes and edges of the thickened polylines
+            var graph = ConstructThickenedPolylineGraph(polylines);
+            // Populate the offset vertex map for each node in the graph. These points can be used to construct the thickened geometry.
+            PopulateOffsetVertexMap(graph, normalDir);
+            var offsetPoints = new List<Vector3>();
+
+            // Map our original vertices to the nodes in the graph. Since we
+            // constructed the graph from these vertices, we can assume that the
+            // nodes are there.
+            var nodeIndicesPerVertex = vertices.Select((v) => graph.PointToNodeIndexMap.GetNearby(v, Vector3.EPSILON).FirstOrDefault()).ToList();
+            for (int i = 0; i < vertices.Count; i++)
+            {
+                var index1 = nodeIndicesPerVertex[i];
+                var index2 = nodeIndicesPerVertex[(i + 1) % vertices.Count];
+                var node1 = graph.Nodes[index1.Value];
+                var offsetVertices = node1.offsetVertexMap[index2.Value];
+                var offsetIndex = left ? 0 : 2;
+                offsetPoints.Add(offsetVertices[offsetIndex]);
+            }
+            return new Polygon(offsetPoints);
         }
     }
 }